Added a temporary action to use with I2C manager retry timer
In my previous version I always just told the timer to re-enter the state listed. For example, if we are in the write data state and we timeout, then re-enter that state again. I had put a note that I was considering adding an action to that callback so that we don't always have to use the "enter" action. Unfortunately, I can't remember right now what it was that I had in mind. I'm sure I'll think of it again soon.