Updated I2C manager error / retry code. Working.
I was forgetting to set the currentDataTransferFinished flag after the error happened. So the manager was not going on to the next device. I added a current error variable also. This way when the error occurs it can be transferred over to the finished error report if needed in case the user needs that information. Now the I2C manager is working properly. When there is no device found, the address gets re-sent multiple times before the error gets set. Then the I2C manager generates a stop condition and goes on to the next device in the list.