Testing bus clear function
The previous version did technically work, but in order to ensure that I make a nice square wave on the SCL line, I needed to keep the SCL pin as an output until I'm completely done. Then set it back to an input. Which means I need one more function for controlling the pin direction.