I2C manager. Working on finished transfer next.
I was having an issue with my pressure sensor. I initially left the pressure sensor unconnected to test the retry timer in the I2C manager. I found a bug in the pressure sensors code, but I also thought of an issue with the way the I2C manager and I2C target currently work. I've sort of been treating them like a UART device. Where when the transfer is finished, I set the finished flag and I'm done. But after working on the error callback and other updates to the I2C manger, I realized that the transfer could finish but with errors. Like when my device isn't connected like right now. The way it's written currently, the data transfer did technically finish, although the device didn't respond. And the error callback did see that there was an error. So it all works. I just haven't decided how I want to handle the transfer fnished event flag. Should it only get set when the transfer was successul?