I2C manager saving progress. Transfer finished code
Working on changing the transfer finished event and flags. One of the issues I've had is that my I2C manager and target code is that if a device times out, we still have to set the finished flag to signal the I2C manager to move on to the next device. This is to keep the I2C manager from getting stuck forever waiting on a device. Technically the device did finish. It just finished with an error. I'm changing this to make it more clear what has happened to the target device.