Updated DataTransfer buffer size code
I'm using the buffer count now instead of the old way which just used the head and tail. In the old method, the buffer needed to be at least one larger than the amount of data you intended to hold. Which wasn't really a big deal for bytes. But this buffer contains complete objects, so I want to be able to reduce any the size in memory if necessary.