I2C manager updates. Added code to catch SCL issue.
Changing the pull up resistor fixes the issue with the SCL clock oscillating. It only happens with the MS5803 pressure sensor because the pressure sensor has very low sink current. The pull up resistors were 1k ohm which is way too low.
Working on seeing if there is a way to recover is software anyway. Simply because the DSPIC really shouldn't being this. It should stop receiving after the 8th bit and then the hardware should clear the RCEN bit. Unfortunately, I can't really do much. All I can do is set the enable bits in software. The hardware is supposed to clear them for me.
The good news is that my I2C machine already has methods in place to catch time out events and do an error callback. When the SCL clock begins malfunctioning, the controller is in the read state. The FsmRead state times out, then the manager calls the FsmStop state, which also times out. I added some code to this last part to reset the I2C module and it does actually work. I am currently testing out this code.
There were a couple times though when the I2C locks up but the SCL line doesn't oscillate. Again, this is all due to the pull up resistors being too low. But I am still trying to come up with different ways to figure out when it happens and catch it in my state machine.