Added an error for when FsmStop times out
Not sure if I'll keep it or not.
I am currently debugging an issue with DSPIC who's I2C bus seems to just start acting crazy every once in a while. The good news is that my I2C manager code already had everything in place to catch the timeout and the error callback function worked perfectly.
While debugging I went ahead and used the "unkown error" in a couple spots temporarily. So in order to tell the difference between the unkown error (IWCOL bit etc.) versus the really bad one that happens when the SCL line just stops working, I added a "flagrant error" in order for me to tell the difference while debugging.