Commented out debug code.
It was interfering with my transmit timeout fail code. The transmit code retries 3 times before calling the error callback. I had thought that the issue with my MS5803 pressure sensor was only happening whenever IWCOL was set so I wrote some code to stop the I2C manager when IWCOL was set. This was not necessary.

There are two different symptoms being shown with this pressure sensor. The first is that the SCL line just gets stuck low. This happens after writing the device's I2C address. SCL gets stuck and the next byte of data does not get sent. Fortunately, sending a stop condition will let it recover, so the I2C manager was already doing a good job on this part. This problem definitely correlates with the pull-up resistor being two low. Originally the I2C bus had 1k resistors and no amount of data could get sent to the pressure sensor.

The second issue is when something happens to the SCL line and it begins to oscillate. I'm still trying to figure what is causing this. It has something to do with the pressure sensor being connected and it seems to also correlate with the pull-up resistor being too low. But I've also seen it on different device when the pull-up resistor value was okay. Either way the DSPIC should never get in this state to begin with so I'm trying to find root causes for both issues.