Updated bus collision code. Ready to test.
Moved it down further with the rest of the errors and removed the line of code that was setting the state of the I2C manager. Now it should no longer halt when there is a bus collision. It should perform the bus clear then keep going to the next device.