Refactored data transfer object in TargetDevice
I decided to go ahead and make the data transfer object required for TargetDevice.h. This is the basic unit that everything is going to be based, whether it ends up being I2C or SPI, or something else. This simplifies the functions a little bit. I also removed the TargetDeviceTransferStatus variable and replaced it with the DataTransfer object.