Reverted target device write to buffer function
It's much easier to use the function this way. Most of time we are using the write function rather than the read function. This is easier to read and should be less prone to mistakes because the user won't have to always update a temporary variable every single time they want to call the function.